refactor: comments-parser
