+++ text/comments-parser/src/extractComments.ts
